Fix POST upload with file as last form field
Previously, if the last field submitted was file contents,
`upload_success` would be marked false even when no errors occur.

This is because the initial value of `upload_success` is false and it is
&'ed with the result of flushing file writers on upload completion.
However, when the last form field is file contents, `upload_success`
would not be set to `true` like it is for other form fields. Therefore,
the value of `upload_success` would always be false.

The solution is to handle the case where the field type is file contents and
set `upload_success` to `true`.