Add support for downloading during upload
When an upload is made via the WebSocket endpoint, it is now possible
for that file to be downloaded while it uploads.

This feature is not added to the POST endpoint because at the moment, to
write an entry to the database we must already have its name and mime
type, but we only get those after the parsing process is completed.

This isn't an impossible obstacle and a possible approach is to move
writing to the database into the function which iterates over the parsed
form contents (that way the database entry can be created as soon as
the required data is available).

However, implementing the above is not as simple as adding support to
the WebSocket endpoint (where the required info is given in the query
string), so it is left out of this commit until I decide how I think it
should be done.