refactor uploading GUI/logic
Make the upload list item appear immediately and ensure error events get
handled, even if they happen *BEFORE* the upload ID is received from the
server.

This resolves the case where the upload fails before the server sends
the ID back to the client which would lead to the upload failing
silently with no visual feedback.