Add is_completed column to uploads table
Add a column to track if an upload is completed. This is used to decide
whether or not to send the file size to download responses. The goal is
that uploads which are being downloaded while they are in progress will
not have a reported size. Previously, the size of the file at the time
the download was started would be reported.

Additionally, a macro is used in db.rs to avoid duplicated match
statements.