Add paste frontend
Add the frontend page for pastes. More work still needs to be done: the
new pages need to be translated.

Also, the paste stuff doesn't currently work without JS and I don't plan
on fixing that, so a warning message should be shown with a link back to
the regular upload/download pages.