Numerous disparate fixes
I went around and tweaked a lot of things. The upload list now has
descriptive error messages if you click the "Failed" indicator. Also,
quotas are changed so that instead of configuring the amount of time
before the quota resets, you give the number of bytes to refund to the
quota each minute.